organize genetic algorithm package a bit
